# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/selection-process": {
      "post": {
        "operationId": "SelectionProcessController_createProcess",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProcessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionProcessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "SelectionProcessController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectionProcessResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/selection-process/interviews": {
      "post": {
        "operationId": "SelectionProcessController_createInterviewSlots",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewSlotsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewSlotResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "SelectionProcessController_findAvailableTimeSlots",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvailableTimeSlotResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Não requer autenticação."
      },
      "patch": {
        "operationId": "SelectionProcessController_bookInterviewSlot",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookInterviewSlotDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewBookingResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Não requer autenticação."
      }
    },
    "/selection-process/interviews/slots": {
      "get": {
        "operationId": "SelectionProcessController_getMySlots",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MySlotResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/selection-process/interviews/send-link": {
      "post": {
        "operationId": "SelectionProcessController_sendInterviewLinks",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendInterviewLinksDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SendLinksResultDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/selection-process/interviews/meet-link": {
      "post": {
        "operationId": "SelectionProcessController_sendMeetLink",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMeetLinkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewBookingResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/selection-process/interviews/{bookingId}/evaluation": {
      "post": {
        "operationId": "SelectionProcessController_createInterviewEvaluation",
        "parameters": [
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewEvaluationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewEvaluationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/selection-process/interviews/evaluations": {
      "get": {
        "operationId": "SelectionProcessController_findInterviewEvaluations",
        "parameters": [
          {
            "name": "selection_process_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewEvaluationWithCandidateResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/selection-process/send-email": {
      "post": {
        "operationId": "SelectionProcessController_sendEmailToCandidates",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendEmailToCandidatesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendEmailResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/selection-process/{processId}": {
      "patch": {
        "operationId": "SelectionProcessController_updateProcess",
        "parameters": [
          {
            "name": "processId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProcessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionProcessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/selection-process/applications": {
      "post": {
        "operationId": "SelectionProcessController_createApplication",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationCreatedResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Não requer autenticação."
      },
      "get": {
        "operationId": "SelectionProcessController_findApplications",
        "parameters": [
          {
            "name": "selection_process_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/selection-process/applications/{applicationId}": {
      "patch": {
        "operationId": "SelectionProcessController_updateApplicationStatus",
        "parameters": [
          {
            "name": "applicationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/selection-process/stages": {
      "post": {
        "operationId": "SelectionProcessController_createStage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "SelectionProcessController_findStages",
        "parameters": [
          {
            "name": "selection_process_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StageResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/selection-process/stages/{stageId}": {
      "put": {
        "operationId": "SelectionProcessController_updateStage",
        "parameters": [
          {
            "name": "stageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/selection-process/candidates": {
      "get": {
        "operationId": "SelectionProcessController_findCandidates",
        "parameters": [
          {
            "name": "selection_process_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/selection-process/candidates/{candidateId}": {
      "patch": {
        "operationId": "SelectionProcessController_updateCandidateStatus",
        "parameters": [
          {
            "name": "candidateId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCandidateStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "SelectionProcess"
        ],
        "description": "Requer autenticação. Restrito a: role: assessor, presidente.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "WattAPI",
    "description": "Documentação gerada a partir dos schemas zod de request/response e do metadata de @RoutePolicy de cada rota.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateProcessDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        },
        "required": [
          "title",
          "starts_at",
          "ends_at"
        ],
        "example": {
          "title": "Processo Seletivo 2026.1",
          "starts_at": "2026-03-01T00:00:00Z",
          "ends_at": "2026-04-01T00:00:00Z"
        }
      },
      "SelectionProcessResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "starts_at": {
            "type": "string"
          },
          "ends_at": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "starts_at",
          "ends_at",
          "created_at"
        ],
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "title": "Processo Seletivo 2026.1",
          "starts_at": "2026-03-01T00:00:00.000Z",
          "ends_at": "2026-04-01T00:00:00.000Z",
          "created_at": "2026-02-15T00:00:00.000Z"
        }
      },
      "CreateInterviewSlotsDto": {
        "type": "object",
        "properties": {
          "slots": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "slots"
        ],
        "example": {
          "slots": [
            "2027-01-20T14:00:00Z",
            "2027-01-21T09:00:00Z"
          ]
        }
      },
      "InterviewSlotResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "selection_process_id": {
            "type": "string"
          },
          "consultant_id": {
            "type": "string"
          },
          "starts_at": {
            "type": "string"
          },
          "ends_at": {
            "type": "string"
          },
          "booking_id": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "selection_process_id",
          "consultant_id",
          "starts_at",
          "ends_at",
          "booking_id",
          "created_at"
        ],
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "selection_process_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "consultant_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "starts_at": "2027-01-20T14:00:00.000Z",
          "ends_at": "2027-01-20T15:00:00.000Z",
          "booking_id": null,
          "created_at": "2026-12-01T00:00:00.000Z"
        }
      },
...
      "SendInterviewLinksDto": {
        "type": "object",
        "properties": {
          "candidate_ids": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
            }
          }
        },
        "required": [
          "candidate_ids"
        ],
        "example": {
          "candidate_ids": [
            "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "5c9d6a2e-8b4f-4e1a-9c3d-7f2a1b6e4d8c"
          ]
        }
      },
...
      "InterviewBookingResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "selection_process_id": {
            "type": "string"
          },
          "candidate_id": {
            "type": "string"
          },
          "starts_at": {
            "type": "string"
          },
          "ends_at": {
            "type": "string"
          },
          "booked_at": {
            "type": "string"
          },
          "meet_link": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "selection_process_id",
          "candidate_id",
          "starts_at",
          "ends_at",
          "booked_at",
          "meet_link",
          "created_at"
        ],
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "selection_process_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "candidate_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "starts_at": "2027-01-01T11:00:00.000Z",
          "ends_at": "2027-01-01T12:00:00.000Z",
          "booked_at": "2026-12-15T00:00:00.000Z",
          "meet_link": "https://meet.google.com/abc-defg-hij",
          "created_at": "2026-12-15T00:00:00.000Z"
        }
      },
...
      "AvailableTimeSlotResponseDto": {
        "type": "object",
        "properties": {
          "starts_at": {
            "type": "string"
          },
          "ends_at": {
            "type": "string"
          }
        },
        "required": [
          "starts_at",
          "ends_at"
        ],
        "example": {
          "starts_at": "2027-01-20T14:00:00.000Z",
          "ends_at": "2027-01-20T15:00:00.000Z"
        }
      },
      "BookInterviewSlotDto": {
        "type": "object",
        "properties": {
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "token": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "starts_at",
          "token"
        ],
        "example": {
          "starts_at": "2027-01-20T14:00:00Z",
          "token": "abc123..."
        }
      },
...
      "SendEmailResultDto": {
        "type": "object",
        "properties": {
          "successes": {
            "type": "number"
          },
          "errors": {
            "type": "number"
          }
        },
        "required": [
          "successes",
          "errors"
        ],
        "example": {
          "successes": 2,
          "errors": 0
        }
      },
...
    }
  }
}